fix-relink modify-client-side

+++ client_bonus.c
@@ -6,7 +6,7 @@
 /*   By: mes-salh <mes-salh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/13 16:28:31 by mes-salh          #+#    #+#             */
-/*   Updated: 2024/02/13 16:28:32 by mes-salh         ###   ########.fr       */
+/*   Updated: 2024/02/25 02:13:37 by mes-salh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,12 +26,17 @@ void	send_bit(int pid, char bit)
 {
 	int	sig;
 
-	if (bit == '0')
-		sig = SIGUSR1;
+	if (pid > 0)
+	{
+		if (bit == '0')
+			sig = SIGUSR1;
+		else
+			sig = SIGUSR2;
+		kill (pid, sig);
+		usleep(150);
+	}
 	else
-		sig = SIGUSR2;
-	kill (pid, sig);
-	usleep(400);
+		return ;
 }
 
 void	send_char(int pid, char c)
@@ -49,6 +54,7 @@ void	send_char(int pid, char c)
 		send_bit (pid, bit);
 		i--;
 	}
+	usleep(150);
 }
 
 int	main(int argc, char **argv)
@@ -62,6 +68,8 @@ int	main(int argc, char **argv)
 		write(2, "Usage: ./client PID message\n", 28);
 		return (1);
 	}
+	if (argv[1][0] == '-')
+		return (0);
 	signal(SIGUSR1, sigal_handler);
 	i = 0;
 	message = argv[2];
